+++ tests/test_filters.cc
@@ -34,7 +34,6 @@ TEST_CASE("Test diagram paths filter", "[unit-test]")
 
     auto cfg = clanguml::config::load("./test_config_data/filters.yml");
 
-    CHECK(cfg.diagrams.size() == 1);
     auto &config = *cfg.diagrams["include_test"];
     clanguml::include_diagram::model::diagram diagram;
 
@@ -51,4 +50,57 @@ TEST_CASE("Test diagram paths filter", "[unit-test]")
     CHECK_FALSE(filter.should_include(make_path("util/error.h")));
     CHECK_FALSE(filter.should_include(
         make_path("sequence_diagram/visitor/translation_unit_visitor.h")));
+}
+
+TEST_CASE("Test method_types include filter", "[unit-test]")
+{
+    using clanguml::class_diagram::model::class_method;
+    using clanguml::common::model::access_t;
+    using clanguml::common::model::diagram_filter;
+    using clanguml::common::model::source_file;
+
+    auto cfg = clanguml::config::load("./test_config_data/filters.yml");
+
+    auto &config = *cfg.diagrams["method_type_include_test"];
+    clanguml::class_diagram::model::diagram diagram;
+
+    diagram_filter filter(diagram, config);
+
+    class_method cm{access_t::kPublic, "A", ""};
+    cm.is_constructor(true);
+
+    CHECK(filter.should_include(cm));
+
+    cm.is_constructor(false);
+    cm.is_destructor(true);
+
+    CHECK(!filter.should_include(cm));
+}
+
+TEST_CASE("Test method_types exclude filter", "[unit-test]")
+{
+    using clanguml::class_diagram::model::class_method;
+    using clanguml::common::model::access_t;
+    using clanguml::common::model::diagram_filter;
+    using clanguml::common::model::source_file;
+
+    auto cfg = clanguml::config::load("./test_config_data/filters.yml");
+
+    auto &config = *cfg.diagrams["method_type_exclude_test"];
+    clanguml::class_diagram::model::diagram diagram;
+
+    diagram_filter filter(diagram, config);
+
+    class_method cm{access_t::kPublic, "A", ""};
+
+    CHECK(filter.should_include(cm));
+
+    cm.is_constructor(true);
+
+    CHECK(filter.should_include(cm));
+
+    cm.is_constructor(false);
+    cm.is_destructor(true);
+
+    CHECK(!filter.should_include(cm));
 }